move unlock to its own 'top-level'
