--- src/webextension/locales/en-US/popup.ftl
@@ -1,7 +1,0 @@
-[[locked]]
-
-locked-title = { product-title }
-locked-tagline = { product-tagline }
-
-locked-signin-action = { product-action-signin }
-locked-prefs-action = { product-action-prefs }
