+++ src/webextension/locales/en-US/unlock.ftl
@@ -1,0 +1,7 @@
+[[unlock]]
+
+unlock-title = { product-title }
+unlock-tagline = { product-tagline }
+
+unlock-action-signin = { product-action-signin }
+unlock-action-prefs = { product-action-prefs }
